medical_backend/requirements.txt
django
djangorestframework
django-cors-headers
google-generativeai
python-dotenv
djangorestframework-simplejwt
